INTYGFV-14680: Added tests for the main converter classes. Also added copyright header for new files.
