+++ af/af00213/src/test/java/se/inera/intyg/common/af00213/v1/model/converter/CertificateToInternalTest.java
@@ -19,28 +19,27 @@
 package se.inera.intyg.common.af00213.v1.model.converter;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.when;
-import static se.inera.intyg.common.af00213.v1.model.converter.certificate.QuestionFunktionsnedsattning.toCertificate;
 
-import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.TestInstance;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import se.inera.intyg.common.af00213.v1.model.converter.certificate.MetaDataGrundData;
 import se.inera.intyg.common.af00213.v1.model.converter.certificate.QuestionAktivitetsbegransning;
 import se.inera.intyg.common.af00213.v1.model.converter.certificate.QuestionArbetspaverkan;
+import se.inera.intyg.common.af00213.v1.model.converter.certificate.QuestionFunktionsnedsattning;
 import se.inera.intyg.common.af00213.v1.model.converter.certificate.QuestionHarAktivitetsbegransning;
 import se.inera.intyg.common.af00213.v1.model.converter.certificate.QuestionHarArbetspaverkan;
+import se.inera.intyg.common.af00213.v1.model.converter.certificate.QuestionHarFunktionsnedsattning;
 import se.inera.intyg.common.af00213.v1.model.converter.certificate.QuestionHarUtredningBehandling;
 import se.inera.intyg.common.af00213.v1.model.converter.certificate.QuestionOvrigt;
 import se.inera.intyg.common.af00213.v1.model.converter.certificate.QuestionUtredningBehandling;
 import se.inera.intyg.common.af00213.v1.model.internal.Af00213UtlatandeV1;
 import se.inera.intyg.common.services.texts.CertificateTextProvider;
 import se.inera.intyg.common.support.facade.builder.CertificateBuilder;
+import se.inera.intyg.common.support.facade.model.Certificate;
 import se.inera.intyg.common.support.model.common.internal.GrundData;
 import se.inera.intyg.common.support.model.common.internal.HoSPersonal;
 import se.inera.intyg.common.support.model.common.internal.Vardenhet;
@@ -49,407 +48,112 @@ import se.inera.intyg.common.support.model.common.internal.Vardenhet;
 class CertificateToInternalTest {
 
     private CertificateTextProvider texts;
+    private Af00213UtlatandeV1 expectedInternalCertificate;
+    private Certificate certificate;
 
     @BeforeEach
     private void setup() {
+        final var grundData = new GrundData();
+        final var hosPersonal = new HoSPersonal();
+        final var vardenhet = new Vardenhet();
+        hosPersonal.setVardenhet(vardenhet);
+        grundData.setSkapadAv(hosPersonal);
+
+        expectedInternalCertificate = Af00213UtlatandeV1.builder()
+            .setGrundData(grundData)
+            .setId("id")
+            .setTextVersion("TextVersion")
+            .setHarFunktionsnedsattning(true)
+            .setFunktionsnedsattning("Funktionsnedsättning")
+            .setHarAktivitetsbegransning(false)
+            .setAktivitetsbegransning("Aktivitetsbegränsning")
+            .setHarUtredningBehandling(true)
+            .setUtredningBehandling("Utredning och behandling")
+            .setHarArbetetsPaverkan(false)
+            .setArbetetsPaverkan("Arbetspåverkan")
+            .setOvrigt("Övrigt")
+            .build();
+
         texts = Mockito.mock(CertificateTextProvider.class);
         when(texts.get(Mockito.any(String.class))).thenReturn("Test string");
+
+        certificate = CertificateBuilder.create()
+            .metadata(MetaDataGrundData.toCertificate(expectedInternalCertificate, texts))
+            .addElement(QuestionHarFunktionsnedsattning.toCertificate(expectedInternalCertificate.getHarFunktionsnedsattning(), 0, texts))
+            .addElement(QuestionFunktionsnedsattning.toCertificate(expectedInternalCertificate.getFunktionsnedsattning(), 0, texts))
+            .addElement(QuestionHarAktivitetsbegransning.toCertificate(expectedInternalCertificate.getHarAktivitetsbegransning(), 0, texts))
+            .addElement(QuestionAktivitetsbegransning.toCertificate(expectedInternalCertificate.getAktivitetsbegransning(), 0, texts))
+            .addElement(QuestionHarUtredningBehandling.toCertificate(expectedInternalCertificate.getHarUtredningBehandling(), 0, texts))
+            .addElement(QuestionUtredningBehandling.toCertificate(expectedInternalCertificate.getUtredningBehandling(), 0, texts))
+            .addElement(QuestionHarArbetspaverkan.toCertificate(expectedInternalCertificate.getHarArbetetsPaverkan(), 0, texts))
+            .addElement(QuestionArbetspaverkan.toCertificate(expectedInternalCertificate.getArbetetsPaverkan(), 0, texts))
+            .addElement(QuestionOvrigt.toCerticate(expectedInternalCertificate.getOvrigt(), 0, texts))
+            .build();
     }
 
-    @Nested
-    @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-    class QuestionFunktionsnedsattningTest {
-
-        private Af00213UtlatandeV1 internalCertificate;
-
-        @BeforeEach
-        void setup() {
-            internalCertificate = Af00213UtlatandeV1.builder()
-                .setGrundData(new GrundData())
-                .setId("id")
-                .setTextVersion("TextVersion")
-                .build();
-        }
-
-        Stream<String> textValues() {
-            return Stream.of("Här kommer en text!", "", null);
-        }
-
-        @ParameterizedTest
-        @MethodSource("textValues")
-        void shouldIncludeFunktionsnedsattningValue(String expectedValue) {
-            final var index = 1;
-
-            final var certificate = CertificateBuilder.create()
-                .addElement(toCertificate(expectedValue, index, texts))
-                .build();
-
-            final var updatedCertificate = CertificateToInternal.convert(certificate, internalCertificate);
-
-            assertEquals(expectedValue, updatedCertificate.getFunktionsnedsattning());
-        }
+    @Test
+    void shallIncludeId() {
+        final var actualInternalCertificate = CertificateToInternal.convert(certificate, expectedInternalCertificate);
+        assertEquals(expectedInternalCertificate.getId(), actualInternalCertificate.getId());
     }
 
-    @Nested
-    @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-    class QuestionHarAktivitetsbegransningTest {
-
-        private Af00213UtlatandeV1 internalCertificate;
-
-        @BeforeEach
-        void setup() {
-            internalCertificate = Af00213UtlatandeV1.builder()
-                .setGrundData(new GrundData())
-                .setId("id")
-                .setTextVersion("TextVersion")
-                .build();
-        }
-
-        Stream<Boolean> booleanValues() {
-            return Stream.of(true, false, null);
-        }
-
-        @ParameterizedTest
-        @MethodSource("booleanValues")
-        void shouldIncludeHarAktivitetsbegransningValue(Boolean expectedValue) {
-            final var index = 1;
-
-            final var certificate = CertificateBuilder.create()
-                .addElement(QuestionHarAktivitetsbegransning.toCertificate(expectedValue, index, texts))
-                .build();
-
-            final var updatedCertificate = CertificateToInternal.convert(certificate, internalCertificate);
-
-            assertEquals(expectedValue, updatedCertificate.getHarAktivitetsbegransning());
-        }
+    @Test
+    void shallIncludeTextVersion() {
+        final var actualInternalCertificate = CertificateToInternal.convert(certificate, expectedInternalCertificate);
+        assertEquals(expectedInternalCertificate.getTextVersion(), actualInternalCertificate.getTextVersion());
     }
 
-    @Nested
-    @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-    class QuestionAktivitetsbegransningTest {
-
-        private Af00213UtlatandeV1 internalCertificate;
-
-        @BeforeEach
-        void setup() {
-            internalCertificate = Af00213UtlatandeV1.builder()
-                .setGrundData(new GrundData())
-                .setId("id")
-                .setTextVersion("TextVersion")
-                .build();
-        }
-
-        Stream<String> textValues() {
-            return Stream.of("Här kommer en text!", "", null);
-        }
-
-        @ParameterizedTest
-        @MethodSource("textValues")
-        void shouldIncludeAktivitetsbegransningValue(String expectedValue) {
-            final var index = 1;
-
-            final var certificate = CertificateBuilder.create()
-                .addElement(QuestionAktivitetsbegransning.toCertificate(expectedValue, index, texts))
-                .build();
-
-            final var updatedCertificate = CertificateToInternal.convert(certificate, internalCertificate);
-
-            assertEquals(expectedValue, updatedCertificate.getAktivitetsbegransning());
-        }
+    @Test
+    void shallIncludeGrundData() {
+        final var actualInternalCertificate = CertificateToInternal.convert(certificate, expectedInternalCertificate);
+        assertNotNull(actualInternalCertificate.getGrundData(), "GrundData is missing!");
     }
 
-    @Nested
-    @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-    class QuestionHarUtredningBehandlingTest {
-
-        private Af00213UtlatandeV1 internalCertificate;
-
-        @BeforeEach
-        void setup() {
-            internalCertificate = Af00213UtlatandeV1.builder()
-                .setGrundData(new GrundData())
-                .setId("id")
-                .setTextVersion("TextVersion")
-                .build();
-        }
-
-        Stream<Boolean> booleanValues() {
-            return Stream.of(true, false, null);
-        }
-
-        @ParameterizedTest
-        @MethodSource("booleanValues")
-        void shouldIncludeHarUtredningBehandlingValue(Boolean expectedValue) {
-            final var index = 1;
-
-            final var certificate = CertificateBuilder.create()
-                .addElement(QuestionHarUtredningBehandling.toCertificate(expectedValue, index, texts))
-                .build();
-
-            final var updatedCertificate = CertificateToInternal.convert(certificate, internalCertificate);
-
-            assertEquals(expectedValue, updatedCertificate.getHarUtredningBehandling());
-        }
+    @Test
+    void shallIncludeHarFunktionsNedsattning() {
+        final var actualInternalCertificate = CertificateToInternal.convert(certificate, expectedInternalCertificate);
+        assertEquals(expectedInternalCertificate.getHarFunktionsnedsattning(), actualInternalCertificate.getHarFunktionsnedsattning());
     }
 
-    @Nested
-    @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-    class QuestionUtredningBehandlingTest {
-
-        private Af00213UtlatandeV1 internalCertificate;
-
-        @BeforeEach
-        void setup() {
-            internalCertificate = Af00213UtlatandeV1.builder()
-                .setGrundData(new GrundData())
-                .setId("id")
-                .setTextVersion("TextVersion")
-                .build();
-        }
-
-        Stream<String> textValues() {
-            return Stream.of("Här kommer en text!", "", null);
-        }
-
-        @ParameterizedTest
-        @MethodSource("textValues")
-        void shouldIncludeUtredningBehandlingValue(String expectedValue) {
-            final var index = 1;
-
-            final var certificate = CertificateBuilder.create()
-                .addElement(QuestionUtredningBehandling.toCertificate(expectedValue, index, texts))
-                .build();
-
-            final var updatedCertificate = CertificateToInternal.convert(certificate, internalCertificate);
-
-            assertEquals(expectedValue, updatedCertificate.getUtredningBehandling());
-        }
+    @Test
+    void shallIncludeFunktionsNedsattning() {
+        final var actualInternalCertificate = CertificateToInternal.convert(certificate, expectedInternalCertificate);
+        assertEquals(expectedInternalCertificate.getFunktionsnedsattning(), actualInternalCertificate.getFunktionsnedsattning());
     }
 
-    @Nested
-    @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-    class QuestionHarArbetetspaverkanTest {
-
-        private Af00213UtlatandeV1 internalCertificate;
-
-        @BeforeEach
-        void setup() {
-            internalCertificate = Af00213UtlatandeV1.builder()
-                .setGrundData(new GrundData())
-                .setId("id")
-                .setTextVersion("TextVersion")
-                .build();
-
-        }
-
-        Stream<Boolean> booleanValues() {
-            return Stream.of(true, false, null);
-        }
-
-        @ParameterizedTest
-        @MethodSource("booleanValues")
-        void shouldIncludeHarArbetetspaverkanValue(Boolean expectedValue) {
-            final var index = 1;
-
-            final var certificate = CertificateBuilder.create()
-                .addElement(QuestionHarArbetspaverkan.toCertificate(expectedValue, index, texts))
-                .build();
-
-            final var updatedCertificate = CertificateToInternal.convert(certificate, internalCertificate);
-
-            assertEquals(expectedValue, updatedCertificate.getHarArbetetsPaverkan());
-        }
+    @Test
+    void shallIncludeHarAktivitetsbegransning() {
+        final var actualInternalCertificate = CertificateToInternal.convert(certificate, expectedInternalCertificate);
+        assertEquals(expectedInternalCertificate.getHarAktivitetsbegransning(), actualInternalCertificate.getHarAktivitetsbegransning());
     }
 
-    @Nested
-    @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-    class QuestionArbetetspaverkanTest {
-
-        private Af00213UtlatandeV1 internalCertificate;
-
-        @BeforeEach
-        void setup() {
-            internalCertificate = Af00213UtlatandeV1.builder()
-                .setGrundData(new GrundData())
-                .setId("id")
-                .setTextVersion("TextVersion")
-                .build();
-        }
-
-        Stream<String> textValues() {
-            return Stream.of("Här kommer en text!", "", null);
-        }
-
-        @ParameterizedTest
-        @MethodSource("textValues")
-        void shouldIncludeArbetetspaverkanValue(String expectedValue) {
-            final var index = 1;
-
-            final var certificate = CertificateBuilder.create()
-                .addElement(QuestionArbetspaverkan.toCertificate(expectedValue, index, texts))
-                .build();
-
-            final var updatedCertificate = CertificateToInternal.convert(certificate, internalCertificate);
-
-            assertEquals(expectedValue, updatedCertificate.getArbetetsPaverkan());
-        }
+    @Test
+    void shallIncludeAktivitetsbegransning() {
+        final var actualInternalCertificate = CertificateToInternal.convert(certificate, expectedInternalCertificate);
+        assertEquals(expectedInternalCertificate.getAktivitetsbegransning(), actualInternalCertificate.getAktivitetsbegransning());
     }
 
-    @Nested
-    @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-    class QuestionOvrigtTest {
-
-        private Af00213UtlatandeV1 internalCertificate;
-
-        @BeforeEach
-        void setup() {
-            internalCertificate = Af00213UtlatandeV1.builder()
-                .setGrundData(new GrundData())
-                .setId("id")
-                .setTextVersion("TextVersion")
-                .build();
-        }
-
-        Stream<String> textValues() {
-            return Stream.of("Här kommer en text!", "", null);
-        }
-
-        @ParameterizedTest
-        @MethodSource("textValues")
-        void shouldIncludeOvrigtValue(String expectedValue) {
-            final var index = 1;
-
-            final var certificate = CertificateBuilder.create()
-                .addElement(QuestionOvrigt.toCerticate(expectedValue, index, texts))
-                .build();
-
-            final var updatedCertificate = CertificateToInternal.convert(certificate, internalCertificate);
-
-            assertEquals(expectedValue, updatedCertificate.getOvrigt());
-        }
+    @Test
+    void shallIncludeHarUtredningBehandling() {
+        final var actualInternalCertificate = CertificateToInternal.convert(certificate, expectedInternalCertificate);
+        assertEquals(expectedInternalCertificate.getHarUtredningBehandling(), actualInternalCertificate.getHarUtredningBehandling());
     }
 
-    @Nested
-    @TestInstance(TestInstance.Lifecycle.PER_CLASS)
-    class GrundDataTestTest {
+    @Test
+    void shallIncludeUtredningBehandling() {
+        final var actualInternalCertificate = CertificateToInternal.convert(certificate, expectedInternalCertificate);
+        assertEquals(expectedInternalCertificate.getUtredningBehandling(), actualInternalCertificate.getUtredningBehandling());
+    }
 
-        private Af00213UtlatandeV1 internalCertificate;
+    @Test
+    void shallIncludeHarArbetetsPaverkan() {
+        final var actualInternalCertificate = CertificateToInternal.convert(certificate, expectedInternalCertificate);
+        assertEquals(expectedInternalCertificate.getHarArbetetsPaverkan(), actualInternalCertificate.getHarArbetetsPaverkan());
+    }
 
-        @BeforeEach
-        void setup() {
-            internalCertificate = Af00213UtlatandeV1.builder()
-                .setGrundData(new GrundData())
-                .setId("id")
-                .setTextVersion("TextVersion")
-                .build();
-        }
-
-        Stream<String> values() {
-            return Stream.of("test string", "", null);
-        }
-
-        @ParameterizedTest
-        @MethodSource("values")
-        void shouldIncludeUnitAddress(String expectedValue) {
-            var grundData = new GrundData();
-            var hosPersonal = new HoSPersonal();
-            var vardenhet = new Vardenhet();
-            vardenhet.setPostadress(expectedValue);
-            hosPersonal.setVardenhet(vardenhet);
-            grundData.setSkapadAv(hosPersonal);
-
-            final var utlatande =
-                Af00213UtlatandeV1.builder()
-                    .setId("id")
-                    .setTextVersion("1.0")
-                    .setGrundData(grundData)
-                    .build();
-
-            final var certificate = CertificateBuilder.create()
-                .metadata(MetaDataGrundData.toCertificate(utlatande, texts))
-                .build();
-
-            final var updatedCertificate = CertificateToInternal.convert(certificate, internalCertificate);
-
-            assertEquals(expectedValue, updatedCertificate.getGrundData().getSkapadAv().getVardenhet().getPostadress());
-        }
-
-        @ParameterizedTest
-        @MethodSource("values")
-        void shouldIncludeUnitCity(String expectedValue) {
-            var grundData = new GrundData();
-            var hosPersonal = new HoSPersonal();
-            var vardenhet = new Vardenhet();
-            vardenhet.setPostort(expectedValue);
-            hosPersonal.setVardenhet(vardenhet);
-            grundData.setSkapadAv(hosPersonal);
-
-            final var utlatande =
-                Af00213UtlatandeV1.builder()
-                    .setId("id")
-                    .setTextVersion("1.0")
-                    .setGrundData(grundData)
-                    .build();
-
-            final var certificate = CertificateBuilder.create()
-                .metadata(MetaDataGrundData.toCertificate(utlatande, texts))
-                .build();
-
-            final var updatedCertificate = CertificateToInternal.convert(certificate, internalCertificate);
-
-            assertEquals(expectedValue, updatedCertificate.getGrundData().getSkapadAv().getVardenhet().getPostort());
-        }
-
-        @ParameterizedTest
-        @MethodSource("values")
-        void shouldIncludeUnitZipCode(String expectedValue) {
-            var grundData = new GrundData();
-            var hosPersonal = new HoSPersonal();
-            var vardenhet = new Vardenhet();
-            vardenhet.setPostnummer(expectedValue);
-            hosPersonal.setVardenhet(vardenhet);
-            grundData.setSkapadAv(hosPersonal);
-
-            final var utlatande =
-                Af00213UtlatandeV1.builder()
-                    .setId("id")
-                    .setTextVersion("1.0")
-                    .setGrundData(grundData)
-                    .build();
-
-            final var certificate = CertificateBuilder.create()
-                .metadata(MetaDataGrundData.toCertificate(utlatande, texts))
-                .build();
-
-            final var updatedCertificate = CertificateToInternal.convert(certificate, internalCertificate);
-
-            assertEquals(expectedValue, updatedCertificate.getGrundData().getSkapadAv().getVardenhet().getPostnummer());
-        }
-
-        @ParameterizedTest
-        @MethodSource("values")
-        void shouldIncludeUnitPhonenumber(String expectedValue) {
-            var grundData = new GrundData();
-            var hosPersonal = new HoSPersonal();
-            var vardenhet = new Vardenhet();
-            vardenhet.setTelefonnummer(expectedValue);
-            hosPersonal.setVardenhet(vardenhet);
-            grundData.setSkapadAv(hosPersonal);
-
-            final var utlatande =
-                Af00213UtlatandeV1.builder()
-                    .setId("id")
-                    .setTextVersion("1.0")
-                    .setGrundData(grundData)
-                    .build();
-
-            final var certificate = CertificateBuilder.create()
-                .metadata(MetaDataGrundData.toCertificate(utlatande, texts))
-                .build();
-
-            final var updatedCertificate = CertificateToInternal.convert(certificate, internalCertificate);
-
-            assertEquals(expectedValue, updatedCertificate.getGrundData().getSkapadAv().getVardenhet().getTelefonnummer());
-        }
+    @Test
+    void shallIncludeArbetetsPaverkan() {
+        final var actualInternalCertificate = CertificateToInternal.convert(certificate, expectedInternalCertificate);
+        assertEquals(expectedInternalCertificate.getArbetetsPaverkan(), actualInternalCertificate.getArbetetsPaverkan());
     }
 }
