+++ af/af00213/src/test/java/se/inera/intyg/common/af00213/v1/model/converter/certificate/CategoryOvrigtTest.java
@@ -1,0 +1,107 @@
+/*
+ * Copyright (C) 2022 Inera AB (http://www.inera.se)
+ *
+ * This file is part of sklintyg (https://github.com/sklintyg).
+ *
+ * sklintyg is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * sklintyg is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package se.inera.intyg.common.af00213.v1.model.converter.certificate;
+
+import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.Mockito.when;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.OVRIGT_CATEGORY_ID;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
+import se.inera.intyg.common.af00213.v1.model.converter.InternalToCertificate;
+import se.inera.intyg.common.af00213.v1.model.internal.Af00213UtlatandeV1;
+import se.inera.intyg.common.services.texts.CertificateTextProvider;
+import se.inera.intyg.common.support.facade.model.config.CertificateDataConfigTypes;
+import se.inera.intyg.common.support.model.common.internal.GrundData;
+import se.inera.intyg.common.support.model.common.internal.HoSPersonal;
+import se.inera.intyg.common.support.model.common.internal.Vardenhet;
+
+class CategoryOvrigtTest {
+
+    private GrundData grundData;
+    private CertificateTextProvider texts;
+
+    @BeforeEach
+    void setup() {
+        final var unit = new Vardenhet();
+
+        final var skapadAv = new HoSPersonal();
+        skapadAv.setVardenhet(unit);
+
+        grundData = new GrundData();
+        grundData.setSkapadAv(skapadAv);
+
+        texts = Mockito.mock(CertificateTextProvider.class);
+        when(texts.get(Mockito.any(String.class))).thenReturn("Test string");
+    }
+
+    @Nested
+    class CategoryOvrigt {
+
+        private Af00213UtlatandeV1 internalCertificate;
+
+        @BeforeEach
+        void createAf00213ToConvert() {
+            internalCertificate = Af00213UtlatandeV1.builder()
+                .setGrundData(grundData)
+                .setId("id")
+                .setTextVersion("TextVersion")
+                .build();
+
+        }
+
+        @Test
+        void shouldIncludeCategoryElement() {
+            final var expectedIndex = 12;
+
+            final var certificate = InternalToCertificate.convert(internalCertificate, texts);
+
+            final var category = certificate.getData().get(OVRIGT_CATEGORY_ID);
+
+            assertAll("Validating category Övrigt",
+                () -> assertEquals(OVRIGT_CATEGORY_ID, category.getId()),
+                () -> assertEquals(expectedIndex, category.getIndex()),
+                () -> assertNull(category.getParent(), "Should not contain a parent"),
+                () -> assertNull(category.getValue(), "Should not contain a value"),
+                () -> assertNull(category.getValidation(), "Should not contain any validation"),
+                () -> assertNotNull(category.getConfig(), "Should include config")
+            );
+        }
+
+        @Test
+        void shouldIncludeCategoryConfig() {
+            final var certificate = InternalToCertificate.convert(internalCertificate, texts);
+
+            final var category = certificate.getData().get(OVRIGT_CATEGORY_ID);
+
+            assertEquals(CertificateDataConfigTypes.CATEGORY, category.getConfig().getType());
+
+            assertAll("Validating category configuration",
+                () -> assertTrue(category.getConfig().getText().trim().length() > 0, "Missing text")
+            );
+        }
+    }
+}
