+++ af/af00213/src/test/java/se/inera/intyg/common/af00213/v1/model/converter/certificate/QuestionOvrigtTest.java
@@ -1,0 +1,200 @@
+/*
+ * Copyright (C) 2022 Inera AB (http://www.inera.se)
+ *
+ * This file is part of sklintyg (https://github.com/sklintyg).
+ *
+ * sklintyg is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * sklintyg is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package se.inera.intyg.common.af00213.v1.model.converter.certificate;
+
+import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.mockito.Mockito.when;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.OVRIGT_CATEGORY_ID;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.OVRIGT_DELSVAR_ID_5;
+import static se.inera.intyg.common.af00213.v1.model.converter.RespConstants.OVRIGT_SVAR_JSON_ID_5;
+
+import java.util.stream.Stream;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+import se.inera.intyg.common.af00213.v1.model.converter.CertificateToInternal;
+import se.inera.intyg.common.af00213.v1.model.converter.InternalToCertificate;
+import se.inera.intyg.common.af00213.v1.model.internal.Af00213UtlatandeV1;
+import se.inera.intyg.common.services.texts.CertificateTextProvider;
+import se.inera.intyg.common.support.facade.builder.CertificateBuilder;
+import se.inera.intyg.common.support.facade.model.config.CertificateDataConfigTextArea;
+import se.inera.intyg.common.support.facade.model.config.CertificateDataConfigTypes;
+import se.inera.intyg.common.support.facade.model.value.CertificateDataTextValue;
+import se.inera.intyg.common.support.model.common.internal.GrundData;
+import se.inera.intyg.common.support.model.common.internal.HoSPersonal;
+import se.inera.intyg.common.support.model.common.internal.Vardenhet;
+
+@ExtendWith(MockitoExtension.class)
+class QuestionOvrigtTest {
+
+    @Mock
+    private CertificateTextProvider texts;
+
+    @BeforeEach
+    void setup() {
+        when(texts.get(Mockito.any(String.class))).thenReturn("Test string");
+    }
+
+    @Nested
+    @TestInstance(TestInstance.Lifecycle.PER_CLASS)
+    class ToInternal {
+
+        private Af00213UtlatandeV1 internalCertificate;
+
+        @BeforeEach
+        void setup() {
+            internalCertificate = Af00213UtlatandeV1.builder()
+                .setGrundData(new GrundData())
+                .setId("id")
+                .setTextVersion("TextVersion")
+                .build();
+        }
+
+        Stream<String> textValues() {
+            return Stream.of("Här kommer en text!", "", null);
+        }
+
+        @ParameterizedTest
+        @MethodSource("textValues")
+        void shouldIncludeOvrigtValue(String expectedValue) {
+            final var index = 1;
+
+            final var certificate = CertificateBuilder.create()
+                .addElement(QuestionOvrigt.toCerticate(expectedValue, index, texts))
+                .build();
+
+            final var updatedCertificate = CertificateToInternal.convert(certificate, internalCertificate);
+
+            assertEquals(expectedValue, updatedCertificate.getOvrigt());
+        }
+    }
+
+    @Nested
+    @TestInstance(Lifecycle.PER_CLASS)
+    class ToCertificate {
+
+        private GrundData grundData;
+
+        @BeforeEach
+        void setup() {
+            final var unit = new Vardenhet();
+
+            final var skapadAv = new HoSPersonal();
+            skapadAv.setVardenhet(unit);
+
+            grundData = new GrundData();
+            grundData.setSkapadAv(skapadAv);
+        }
+
+        @Nested
+        class QuestionOvrigt {
+
+            private Af00213UtlatandeV1 internalCertificate;
+
+            @BeforeEach
+            void createAf00213ToConvert() {
+                internalCertificate = Af00213UtlatandeV1.builder()
+                    .setGrundData(grundData)
+                    .setId("id")
+                    .setTextVersion("TextVersion")
+                    .setOvrigt("Text som beskriver Övrigt")
+                    .build();
+            }
+
+            @Test
+            void shouldIncludeQuestionElement() {
+                final var expectedIndex = 13;
+
+                final var certificate = InternalToCertificate.convert(internalCertificate, texts);
+
+                final var question = certificate.getData().get(OVRIGT_DELSVAR_ID_5);
+
+                assertAll("Validating question Övrigt",
+                    () -> assertEquals(OVRIGT_DELSVAR_ID_5, question.getId()),
+                    () -> assertEquals(expectedIndex, question.getIndex()),
+                    () -> assertEquals(OVRIGT_SVAR_JSON_ID_5, question.getParent()),
+                    () -> assertNotNull(question.getValue(), "Should include a value"),
+                    () -> assertNull(question.getValidation(), "Shouldn't include validation"),
+                    () -> assertNotNull(question.getConfig(), "Should include config")
+                );
+            }
+
+            @Test
+            void shouldIncludeQuestionConfig() {
+                final var certificate = InternalToCertificate.convert(internalCertificate, texts);
+
+                final var question = certificate.getData().get(OVRIGT_DELSVAR_ID_5);
+
+                assertEquals(CertificateDataConfigTypes.UE_TEXTAREA, question.getConfig().getType());
+
+                final var certificateDataConfigTextArea = (CertificateDataConfigTextArea) question.getConfig();
+                assertAll("Validating question configuration",
+                    () -> assertTrue(certificateDataConfigTextArea.getText().trim().length() > 0, "Missing text"),
+                    () -> assertTrue(certificateDataConfigTextArea.getDescription().trim().length() > 0, "Missing description"),
+                    () -> assertEquals(OVRIGT_SVAR_JSON_ID_5, certificateDataConfigTextArea.getId())
+                );
+            }
+
+            @Test
+            void shouldIncludeQuestionValue() {
+                final var certificate = InternalToCertificate.convert(internalCertificate, texts);
+
+                final var question = certificate.getData().get(OVRIGT_DELSVAR_ID_5);
+
+                final var certificateDataTextValue = (CertificateDataTextValue) question.getValue();
+                assertAll("Validating question value",
+                    () -> assertEquals(OVRIGT_CATEGORY_ID, certificateDataTextValue.getId()),
+                    () -> assertEquals(internalCertificate.getOvrigt(), certificateDataTextValue.getText())
+                );
+            }
+
+            @Test
+            void shouldIncludeQuestionValueEmpty() {
+                internalCertificate = Af00213UtlatandeV1.builder()
+                    .setGrundData(grundData)
+                    .setId("id")
+                    .setTextVersion("TextVersion")
+                    .build();
+
+                final var certificate = InternalToCertificate.convert(internalCertificate, texts);
+
+                final var question = certificate.getData().get(OVRIGT_DELSVAR_ID_5);
+
+                final var certificateDataTextValue = (CertificateDataTextValue) question.getValue();
+                assertAll("Validating question value",
+                    () -> assertEquals(OVRIGT_SVAR_JSON_ID_5, certificateDataTextValue.getId()),
+                    () -> assertNull(certificateDataTextValue.getText())
+                );
+            }
+        }
+    }
+}
